Fix EventHorizon int properties expecting Long
The generated SyncedTranscripts*Event classes type schema int properties as
Long. Use Long expressions for from/to_position_seconds, manual_scroll_duration_ms,
episode_duration_seconds and duration_ms.

+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/fingerprint/FingerprintTimingManager.kt
@@ -193,8 +193,8 @@ class FingerprintTimingManager @Inject constructor(
         Timber.d("FingerprintTimingManager: stopped")
     }
 
-    private fun elapsedPreparationMs(): Int =
-        if (preparationStartMs == 0L) 0 else (System.currentTimeMillis() - preparationStartMs).toInt()
+    private fun elapsedPreparationMs(): Long =
+        if (preparationStartMs == 0L) 0L else System.currentTimeMillis() - preparationStartMs
 
     private fun markUnavailable(
         reason: String,
@@ -318,7 +318,7 @@ class FingerprintTimingManager @Inject constructor(
         preparationStartMs = System.currentTimeMillis()
         eventHorizon.track(
             SyncedTranscriptsPreparationStartedEvent(
-                episodeDurationSeconds = durationSec.toInt(),
+                episodeDurationSeconds = durationSec.toLong(),
                 isStreaming = currentIsStreaming,
                 episodeUuid = episodeUuid,
             ),
